use smart pointers and fix some warnings

+++ algorithms/0001_two_sum.cpp
@@ -38,12 +38,12 @@ class Solution
     std::vector<int> twoSumBrute(const std::vector<int>& nums, int target)
     {
 
-        for (int i = 0; i < nums.size(); i++)
+        for (size_t i = 0; i < nums.size(); i++)
         {
-            for (int j = i + 1; j < nums.size(); j++)
+            for (size_t j = i + 1; j < nums.size(); j++)
             {
                 if (nums[i] + nums[j] == target)
-                    return std::vector{i, j};
+                    return std::vector{static_cast<int>(i), static_cast<int>(j)};
             }
         }
